+++ src/com/company/ProductFactory.java
@@ -4,8 +4,8 @@ public interface ProductFactory {
 
     public CPU buildCpu();
     public Screen buildScreen();
-    void setCPUPrice(CPU cpu,  float price);
-    void setScreenPrice(Screen screen, float price);
+    void setCPUPrice(CPU cpu,  double price);
+    void setScreenPrice(Screen screen, double price);
     void setScreenName(Screen screen, String name);
     void setCPUName(CPU cpu, String name);
 }
